requirements.txt
Django>=4.2
djangorestframework>=3.15
djangorestframework-simplejwt>=5.2.2
drf-yasg>=1.21.6
PyYAML
psycopg2-binary>=2.9